move FIFO entry-point tests into run_command.test.ts
