+++ test/cli/run/run_command.test.ts
@@ -1,7 +1,10 @@
 import { spawnSync } from "bun";
 import { describe, expect, test } from "bun:test";
 import { rmSync, writeFileSync } from "fs";
-import { bunEnv, bunExe, bunRun, isWindows, tempDir } from "harness";
+import { bunEnv, bunExe, bunRun, isPosix, isWindows, tempDir } from "harness";
+import { mkfifo } from "mkfifo";
+import { open } from "node:fs/promises";
+import { join } from "path";
 
 let cwd: string;
 
@@ -36,6 +39,64 @@ describe("bun", () => {
     });
     expect(stdout.toString()).toBeEmpty();
     expect(stderr.toString()).toMatch(/Script not found/);
+    expect(exitCode).toBe(1);
+  });
+});
+
+// #7102: `bun <(cmd)` hands bun a /dev/fd/N path that opens to a pipe. The CLI
+// used to readlink that to `pipe:[inode]` (Linux) or fail F_GETPATH (macOS)
+// and then try to load the module from that non-path. A named FIFO is the same
+// shape without a bash dependency.
+describe.skipIf(!isPosix)("entry point is a pipe/FIFO", () => {
+  test.concurrent("runs a named FIFO entry point", async () => {
+    using dir = tempDir("run-fifo", {});
+    const fifo = join(String(dir), "script");
+    mkfifo(fifo);
+
+    await using proc = Bun.spawn({
+      cmd: [bunExe(), fifo, "extra"],
+      env: bunEnv,
+      stdio: ["ignore", "pipe", "pipe"],
+    });
+    const feed = (async () => {
+      const w = await open(fifo, "w");
+      await w.writeFile(`console.log(JSON.stringify({ argv: process.argv.slice(1), main: import.meta.main }));\n`);
+      await w.close();
+    })().catch(e => e);
+
+    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
+    await feed;
+    expect(stderr).toBe("");
+    expect(JSON.parse(stdout)).toEqual({ argv: [fifo, "extra"], main: true });
+    expect(exitCode).toBe(0);
+  });
+
+  const bash = Bun.which("bash");
+  test.concurrent.skipIf(!bash)("runs a process-substitution entry point (bash <())", async () => {
+    const script = `console.log(JSON.stringify({ argv1: process.argv[1], main: import.meta.main, sum: 1 + 2 }))`;
+    await using proc = Bun.spawn({
+      cmd: [bash!, "-c", `exec "$BUN" <(printf %s "$SCRIPT")`],
+      env: { ...bunEnv, BUN: bunExe(), SCRIPT: script },
+      stdio: ["ignore", "pipe", "pipe"],
+    });
+    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
+    expect(stderr).toBe("");
+    const parsed = JSON.parse(stdout);
+    expect(parsed.argv1).toMatch(/^\/dev\/fd\/\d+$/);
+    expect(parsed).toEqual({ argv1: parsed.argv1, main: true, sum: 3 });
+    expect(exitCode).toBe(0);
+  });
+
+  test.concurrent.skipIf(!bash)("error output references a path, not 'pipe:[inode]'", async () => {
+    await using proc = Bun.spawn({
+      cmd: [bash!, "-c", `exec "$BUN" <(printf %s 'throw new Error("boom")')`],
+      env: { ...bunEnv, BUN: bunExe() },
+      stdio: ["ignore", "pipe", "pipe"],
+    });
+    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
+    expect(stdout).toBe("");
+    expect(stderr).toContain("boom");
+    expect(stderr).not.toMatch(/pipe:\[\d+\]/);
     expect(exitCode).toBe(1);
   });
 });
